refactor: 중복 코드 제거 및 Redirect View변화로 인한 테스트 수정
