+++ src/main/resources/application.properties
@@ -11,3 +11,4 @@ spring.datasource.password=1234
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 
+server.servlet.session.tracking-modes=cookie
